+++ src/main/java/ma/dentaire/projetdentaires8/service/IServiceFacture.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 public interface IServiceFacture {
     List<FactureShowDto> findFacture();
-    public List<FactureShowDto> findFacturebyPatient(int patientId);
+    public List<FactureShowDto> findFacturebyPatient(Long patientId);
     Facture addFacture(FactureAddDto facture);
     Facture updateFacture(FactureUpdateDto factureUpdate);
     void deleteFacture(int factureId);
